convert legacy AnnotatedFunction to Action
