+++ CopilotKit/packages/backend/src/lib/copilotkit-backend.ts
@@ -1,5 +1,6 @@
 import {
   Action,
+  AnnotatedFunction,
   Parameter,
 } from "@copilotkit/shared";
 import { CopilotBackendImplementation } from "./copilotkit-backend-implementation";
@@ -11,9 +12,18 @@ interface CopilotBackendConstructorParams<T extends Parameter[]| [] = []> {
   debug?: boolean;
 }
 
+interface CopilotDeprecatedBackendConstructorParams<T extends Parameter[]| [] = []> {
+  actions?: AnnotatedFunction<any>[];
+  langserve?: RemoteChain[];
+  debug?: boolean;
+}
+
 
 export class CopilotBackend<const T extends Parameter[]| [] = []> extends CopilotBackendImplementation {
-  constructor(params?: CopilotBackendConstructorParams<T>) {
+  constructor(params?: CopilotBackendConstructorParams<T>);
+  // @deprecated use Action<T> instead of AnnotatedFunction<T>
+  constructor(params?: CopilotDeprecatedBackendConstructorParams<T>);
+  constructor(params?: CopilotBackendConstructorParams<T> | CopilotDeprecatedBackendConstructorParams<T>) {
     super(params);
   }
 
@@ -21,7 +31,10 @@ export class CopilotBackend<const T extends Parameter[]| [] = []> extends Copilo
   // To have the main implementation checked by prettier, we split 
   // this into a separate file
   // prettier-ignore
-  addAction<const T extends Parameter[] | [] = []>(action: Action<T>): void {
+  addAction<const T extends Parameter[] | [] = []>(action: Action<T>): void;
+  /** @deprecated Use addAction with Action<T> instead. */
+  addAction(action: AnnotatedFunction<any>): void;
+  addAction<const T extends Parameter[] | [] = []>(action: Action<T> | AnnotatedFunction<any>): void {
     super.addAction(action);
   }
 }
