+++ CopilotKit/packages/react-core/src/hooks/use-copilot-action-implementation.ts
@@ -1,6 +1,6 @@
 "use client";
 
-import { AnnotatedFunction, Parameter } from "@copilotkit/shared";
+import { Action, AnnotatedFunction, Parameter } from "@copilotkit/shared";
 import { useRef, useContext, useEffect } from "react";
 import { FrontendAction } from "../types/frontend-action";
 import { CopilotContext } from "../context/copilot-context";
@@ -60,55 +60,4 @@ export function useCopilotActionImplementation<T extends Array<any> = []>(
     // dependencies set by the developer
     ...(dependencies || []),
   ]);
-}
-
-export function annotatedFunctionToAction(
-  annotatedFunction: AnnotatedFunction<any[]>,
-): FrontendAction<any> {
-  const parameters: Parameter[] = annotatedFunction.argumentAnnotations.map((annotation) => {
-    switch (annotation.type) {
-      case "string":
-      case "number":
-      case "boolean":
-      case "object":
-        return {
-          name: annotation.name,
-          description: annotation.description,
-          type: annotation.type,
-          required: annotation.required,
-        };
-      case "array":
-        let type;
-        if (annotation.items.type === "string") {
-          type = "string[]";
-        } else if (annotation.items.type === "number") {
-          type = "number[]";
-        } else if (annotation.items.type === "boolean") {
-          type = "boolean[]";
-        } else if (annotation.items.type === "object") {
-          type = "object[]";
-        } else {
-          type = "string[]";
-        }
-        return {
-          name: annotation.name,
-          description: annotation.description,
-          type: type as any,
-          required: annotation.required,
-        };
-    }
-  });
-
-  return {
-    name: annotatedFunction.name,
-    description: annotatedFunction.description,
-    parameters: parameters,
-    handler: (args) => {
-      const paramsInCorrectOrder: any[] = [];
-      for (let arg of annotatedFunction.argumentAnnotations) {
-        paramsInCorrectOrder.push(args[arg.name]);
-      }
-      return annotatedFunction.implementation(...paramsInCorrectOrder);
-    },
-  };
 }
